refactor(analysis): move ui segment + add public API barrel
AnalysisCTA/AnalysisCTAContent and the useAnalysis composition-root hook -> ui/.
Add index.ts as the slice's public API; rewrite the README for the FSD layout
and drop the obsolete per-ring READMEs.
