+++ src/features/analysis/ui/__tests__/use-analysis.test.tsx
@@ -1,9 +1,9 @@
 // @vitest-environment happy-dom
 import { act, renderHook, waitFor } from "@testing-library/react";
 import { describe, expect, it, vi } from "vitest";
-import type { AnalysisService } from "../../application/analysis-service";
-import type { AnalysisSelection } from "../../domain/analysis-selection";
-import type { InsightSink } from "../insight-sink";
+import type { AnalysisService } from "../../model/analysis-service";
+import type { AnalysisSelection } from "../../model/analysis-selection";
+import type { InsightSink } from "../../model/insight-sink";
 import { useAnalysis } from "../use-analysis";
 
 const selection: AnalysisSelection = {
